refactor: changed focusGallery file to index.jsx as it may have been mistaken for a route as it was named page.jsx
